+++ Chewbacca.sln
@@ -3,26 +3,15 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.0.31903.59
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WebApi", "src\WebApi\WebApi.csproj", "{FCD78AF4-ABC6-44BC-90EB-DEF5DC725434}"
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{D4301FB6-9DC3-4C6F-9A75-D43C917AE580}"
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{2CBAF712-D99B-417D-8465-F2AA9E99D818}"
-	ProjectSection(SolutionItems) = preProject
-		src\CvPartnerService\obj\project.assets.json = src\CvPartnerService\obj\project.assets.json
-		src\CvPartnerService\obj\project.nuget.cache = src\CvPartnerService\obj\project.nuget.cache
-		src\CvPartnerService\Properties\launchSettings.json = src\CvPartnerService\Properties\launchSettings.json
-	EndProjectSection
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CvPartner", "src\CvPartner\CvPartner.csproj", "{D0E1CE79-F0BE-4268-AF66-388B176CC540}"
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "tests", "tests", "{9E85E787-8541-4B07-BF02-743E9B864D58}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Shared", "src\Shared\Shared.csproj", "{D164C471-C992-477C-AB72-61A22FB3F863}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "IntegrationTests", "tests\IntegrationTests\IntegrationTests.csproj", "{D0D3B66A-F29D-4A16-853F-1FD355E87F5D}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WebApi", "src\WebApi\WebApi.csproj", "{737A6CBC-94B0-481A-B0FD-3266B57B9839}"
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Repositories", "Repositories", "{4F23BE32-4E92-4E49-95D1-9DA250B8876E}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CvPartner", "src\CvPartner\CvPartner.csproj", "{9A5AE8ED-23F4-488F-8D6B-619712F793F0}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Employee", "src\Employee\Employee.csproj", "{EA67E5E8-F741-4180-AD73-ADC9EC165363}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Shared", "src\Shared\Shared.csproj", "{3E96FF34-38BE-4BED-ACBA-0796FBD2AB70}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Employee", "src\Employee\Employee.csproj", "{931027B4-5F64-4849-8048-0EA5CB4FA419}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -33,33 +22,27 @@ Global
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{FCD78AF4-ABC6-44BC-90EB-DEF5DC725434}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{FCD78AF4-ABC6-44BC-90EB-DEF5DC725434}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{FCD78AF4-ABC6-44BC-90EB-DEF5DC725434}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{FCD78AF4-ABC6-44BC-90EB-DEF5DC725434}.Release|Any CPU.Build.0 = Release|Any CPU
-		{D0D3B66A-F29D-4A16-853F-1FD355E87F5D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D0D3B66A-F29D-4A16-853F-1FD355E87F5D}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{D0D3B66A-F29D-4A16-853F-1FD355E87F5D}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D0D3B66A-F29D-4A16-853F-1FD355E87F5D}.Release|Any CPU.Build.0 = Release|Any CPU
-		{9A5AE8ED-23F4-488F-8D6B-619712F793F0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{9A5AE8ED-23F4-488F-8D6B-619712F793F0}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{9A5AE8ED-23F4-488F-8D6B-619712F793F0}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{9A5AE8ED-23F4-488F-8D6B-619712F793F0}.Release|Any CPU.Build.0 = Release|Any CPU
-		{EA67E5E8-F741-4180-AD73-ADC9EC165363}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{EA67E5E8-F741-4180-AD73-ADC9EC165363}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{EA67E5E8-F741-4180-AD73-ADC9EC165363}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{EA67E5E8-F741-4180-AD73-ADC9EC165363}.Release|Any CPU.Build.0 = Release|Any CPU
-		{3E96FF34-38BE-4BED-ACBA-0796FBD2AB70}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{3E96FF34-38BE-4BED-ACBA-0796FBD2AB70}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{3E96FF34-38BE-4BED-ACBA-0796FBD2AB70}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{3E96FF34-38BE-4BED-ACBA-0796FBD2AB70}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D0E1CE79-F0BE-4268-AF66-388B176CC540}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D0E1CE79-F0BE-4268-AF66-388B176CC540}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D0E1CE79-F0BE-4268-AF66-388B176CC540}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D0E1CE79-F0BE-4268-AF66-388B176CC540}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D164C471-C992-477C-AB72-61A22FB3F863}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D164C471-C992-477C-AB72-61A22FB3F863}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D164C471-C992-477C-AB72-61A22FB3F863}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D164C471-C992-477C-AB72-61A22FB3F863}.Release|Any CPU.Build.0 = Release|Any CPU
+		{737A6CBC-94B0-481A-B0FD-3266B57B9839}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{737A6CBC-94B0-481A-B0FD-3266B57B9839}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{737A6CBC-94B0-481A-B0FD-3266B57B9839}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{737A6CBC-94B0-481A-B0FD-3266B57B9839}.Release|Any CPU.Build.0 = Release|Any CPU
+		{931027B4-5F64-4849-8048-0EA5CB4FA419}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{931027B4-5F64-4849-8048-0EA5CB4FA419}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{931027B4-5F64-4849-8048-0EA5CB4FA419}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{931027B4-5F64-4849-8048-0EA5CB4FA419}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
-		{FCD78AF4-ABC6-44BC-90EB-DEF5DC725434} = {2CBAF712-D99B-417D-8465-F2AA9E99D818}
-		{D0D3B66A-F29D-4A16-853F-1FD355E87F5D} = {9E85E787-8541-4B07-BF02-743E9B864D58}
-		{4F23BE32-4E92-4E49-95D1-9DA250B8876E} = {2CBAF712-D99B-417D-8465-F2AA9E99D818}
-		{9A5AE8ED-23F4-488F-8D6B-619712F793F0} = {4F23BE32-4E92-4E49-95D1-9DA250B8876E}
-		{EA67E5E8-F741-4180-AD73-ADC9EC165363} = {2CBAF712-D99B-417D-8465-F2AA9E99D818}
-		{3E96FF34-38BE-4BED-ACBA-0796FBD2AB70} = {2CBAF712-D99B-417D-8465-F2AA9E99D818}
+		{D0E1CE79-F0BE-4268-AF66-388B176CC540} = {D4301FB6-9DC3-4C6F-9A75-D43C917AE580}
+		{D164C471-C992-477C-AB72-61A22FB3F863} = {D4301FB6-9DC3-4C6F-9A75-D43C917AE580}
+		{737A6CBC-94B0-481A-B0FD-3266B57B9839} = {D4301FB6-9DC3-4C6F-9A75-D43C917AE580}
+		{931027B4-5F64-4849-8048-0EA5CB4FA419} = {D4301FB6-9DC3-4C6F-9A75-D43C917AE580}
 	EndGlobalSection
 EndGlobal
