+++ src/CvPartner/DTOs.cs
