+++ src/Shared/AppSettings.cs
@@ -1,4 +1,6 @@
-﻿namespace Shared;
+﻿using System;
+
+namespace Shared;
 
 public record AppSettings
 {
