+++ src/WebApi/WebApi.csproj
@@ -22,6 +22,7 @@
   <ItemGroup>
     <ProjectReference Include="..\Employee\Employee.csproj" />
     <ProjectReference Include="..\Shared\Shared.csproj" />
+    <ProjectReference Include="..\CvPartner\CvPartner.csproj" />
   </ItemGroup>
     
   <ItemGroup>
